+++ NEWS.md
@@ -1,3 +1,5 @@
+# distanceto (development version)
+
 # distanceto 0.0.1
 
 * initial 0.0.1 with `distance_to()` and `distance_raster()`