+++ NEWS.md
@@ -1,5 +1,9 @@
 # distanceto (development version)
 
+* add basic set of tests with [`tinytest`](https://github.com/markvanderloo/tinytest/)
+* use GitHub Actions for test workflow
+* fix distance calculation when crs indicates long lat, unprojected coordinates
+
 # distanceto 0.0.1
 
 * initial 0.0.1 with `distance_to()` and `distance_raster()`