+++ NEWS.md
@@ -1,5 +1,8 @@
 # distanceto (development version)
 
+* add "measure" argument to `distance_raster`
+* fill `distance_raster` man
+
 # distanceto 0.0.1
 
 * initial 0.0.1 with `distance_to()` and `distance_raster()`