+++ NEWS.md
@@ -1,7 +1,11 @@
 # distanceto (development version)
 
-* add "measure" argument to `distance_raster`
+* add basic set of tests with [`tinytest`](https://github.com/markvanderloo/tinytest/)
 * fill `distance_raster` man
+* fix distance calculation when crs indicates long lat, unprojected coordinates
+* use GitHub Actions for test workflow
+* add "measure" argument to `distance_raster`
+
 
 # distanceto 0.0.1