+++ NEWS.md
@@ -1,5 +1,12 @@
 # distanceto (development version)
 
+* add basic set of tests for `distance_to` and `distance_raster` with [`tinytest`](https://github.com/markvanderloo/tinytest/)
+* fill `distance_raster` man
+* fix distance calculation when crs indicates long lat, unprojected coordinates
+* use GitHub Actions for test workflow
+* add "measure" argument to `distance_raster`
+
+
 # distanceto 0.0.1
 
 * initial 0.0.1 with `distance_to()` and `distance_raster()`